Update pool ID type and enhance PoolInfo management
Changed the pool ID type from usize to u32 for better consistency across the project. Additionally, restructured PoolInfoList to include a HashMap for efficient access and implemented related methods. Updated graph path estimation to use the new structure and simplified return handling by replacing BigRational with u128.

+++ src/ref_finance/errors.rs
@@ -8,7 +8,7 @@ pub enum Error {
     SwapSameToken,
     ZeroAmount,
     Overflow,
-    OutOfIndexOfPools(usize),
+    OutOfIndexOfPools(u32),
     OutOfIndexOfTokens(TokenIndex),
     DifferentLengthOfTokens(usize, usize),
     InvalidPoolSize(usize),
